+++ api/tests/routes/public/individual_offers/v1/endpoints_test.py
@@ -11,7 +11,6 @@ from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offers import factories as offers_factories
 from pcapi.core.offers import models as offers_models
-from pcapi.utils import human_ids
 
 from . import utils
 
@@ -172,257 +171,6 @@ class GetEventTest:
 
         assert response.status_code == 200
         assert response.json["ticketCollection"] is None
-
-
-@pytest.mark.usefixtures("db_session")
-class GetEventDatesTest:
-    @freezegun.freeze_time("2023-01-01 12:00:00")
-    def test_event_with_dates(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(venue=venue)
-        offers_factories.StockFactory(offer=event_offer, isSoftDeleted=True)
-        price_category = offers_factories.PriceCategoryFactory(
-            offer=event_offer, price=12.34, priceCategoryLabel__label="carre or"
-        )
-        bookable_stock = offers_factories.EventStockFactory(
-            offer=event_offer,
-            priceCategory=price_category,
-            quantity=10,
-            bookingLimitDatetime=datetime.datetime(2023, 1, 15, 13, 0, 0),
-            beginningDatetime=datetime.datetime(2023, 1, 15, 13, 0, 0),
-        )
-        stock_without_booking = offers_factories.EventStockFactory(
-            offer=event_offer,
-            # FIXME (cepehang, 2023-02-02): remove price and None price category after price category generation script
-            price=12.34,
-            priceCategory=None,
-            quantity=2,
-            bookingLimitDatetime=datetime.datetime(2023, 1, 15, 13, 0, 0),
-            beginningDatetime=datetime.datetime(2023, 1, 15, 13, 0, 0),
-        )
-        offers_factories.EventStockFactory(offer=event_offer, isSoftDeleted=True)  # deleted stock, not returned
-        bookings_factories.BookingFactory(stock=bookable_stock)
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/events/{event_offer.id}/dates"
-            )
-
-        assert response.status_code == 200
-        assert response.json["dates"] == [
-            {
-                "beginningDatetime": "2023-01-15T13:00:00Z",
-                "bookedQuantity": 1,
-                "bookingLimitDatetime": "2023-01-15T13:00:00Z",
-                "id": bookable_stock.id,
-                "priceCategory": {"id": price_category.id, "label": "carre or", "price": 1234},
-                "quantity": 10,
-            },
-            {
-                "beginningDatetime": "2023-01-15T13:00:00Z",
-                "bookedQuantity": 0,
-                "bookingLimitDatetime": "2023-01-15T13:00:00Z",
-                "id": stock_without_booking.id,
-                "priceCategory": {"id": None, "label": None, "price": 1234},
-                "quantity": 2,
-            },
-        ]
-        assert (
-            response.json["pagination"]["pagesLinks"]["current"]
-            == f"http://localhost/public/offers/v1/events/{event_offer.id}/dates?page=1&limit=50"
-        )
-
-    def test_event_without_dates(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(venue=venue)
-        offers_factories.EventStockFactory(offer=event_offer, isSoftDeleted=True)  # deleted stock, not returned
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/events/{event_offer.id}/dates"
-            )
-
-        assert response.status_code == 200
-        assert response.json == {
-            "dates": [],
-            "pagination": {
-                "currentPage": 1,
-                "itemsCount": 0,
-                "itemsTotal": 0,
-                "lastPage": 1,
-                "limitPerPage": 50,
-                "pagesLinks": {
-                    "current": f"http://localhost/public/offers/v1/events/{event_offer.id}/dates?page=1&limit=50",
-                    "first": f"http://localhost/public/offers/v1/events/{event_offer.id}/dates?page=1&limit=50",
-                    "last": f"http://localhost/public/offers/v1/events/{event_offer.id}/dates?page=1&limit=50",
-                    "next": None,
-                    "previous": None,
-                },
-            },
-        }
-
-    def test_404_when_page_is_too_high(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        event_offer = offers_factories.EventOfferFactory(venue=venue)
-        offers_factories.EventStockFactory(offer=event_offer)  # deleted stock, not returned
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/events/{event_offer.id}/dates?page=2&limit=50"
-            )
-
-        assert response.status_code == 404
-        assert response.json == {
-            "page": "The page you requested does not exist. The maximum page for the specified limit is 1"
-        }
-
-
-@pytest.mark.usefixtures("db_session")
-class GetProductsTest:
-    ENDPOINT_URL = "http://localhost/public/offers/v1/products"
-
-    def test_get_first_page(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        offers = offers_factories.ThingOfferFactory.create_batch(12, venue=venue)
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/products?venueId={venue.id}&limit=5"
-            )
-
-        assert response.status_code == 200
-        assert response.json["pagination"] == {
-            "currentPage": 1,
-            "itemsCount": 5,
-            "itemsTotal": 12,
-            "lastPage": 3,
-            "limitPerPage": 5,
-            "pagesLinks": {
-                "current": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
-                "first": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
-                "last": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=3&limit=5",
-                "next": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=2&limit=5",
-                "previous": None,
-            },
-        }
-        assert [product["id"] for product in response.json["products"]] == [offer.id for offer in offers[0:5]]
-
-    def test_get_last_page(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        offers = offers_factories.ThingOfferFactory.create_batch(12, venue=venue)
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/products?venueId={venue.id}&limit=5&page=3"
-            )
-
-        assert response.status_code == 200
-        assert response.json["pagination"] == {
-            "currentPage": 3,
-            "itemsCount": 2,
-            "itemsTotal": 12,
-            "lastPage": 3,
-            "limitPerPage": 5,
-            "pagesLinks": {
-                "current": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=3&limit=5",
-                "first": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
-                "last": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=3&limit=5",
-                "next": None,
-                "previous": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=2&limit=5",
-            },
-        }
-        assert [product["id"] for product in response.json["products"]] == [offer.id for offer in offers[10:12]]
-
-    def test_404_when_the_page_is_too_high(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/products?venueId={venue.id}&limit=5&page=2"
-            )
-        assert response.status_code == 404
-        assert response.json == {
-            "page": "The page you requested does not exist. The maximum page for the " "specified limit is 1"
-        }
-
-    def test_200_for_first_page_if_no_items(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/products?venueId={venue.id}&limit=5"
-            )
-
-        assert response.status_code == 200
-        assert response.json == {
-            "pagination": {
-                "currentPage": 1,
-                "itemsCount": 0,
-                "itemsTotal": 0,
-                "limitPerPage": 5,
-                "lastPage": 1,
-                "pagesLinks": {
-                    "current": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
-                    "first": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
-                    "last": f"{self.ENDPOINT_URL}?venueId={venue.id}&page=1&limit=5",
-                    "next": None,
-                    "previous": None,
-                },
-            },
-            "products": [],
-        }
-
-    def test_400_when_limit_is_too_high(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/products?venueId={venue.id}&limit=51"
-            )
-
-        assert response.status_code == 400
-        assert response.json == {"limit": ["ensure this value is less than or equal to 50"]}
-
-    def test_get_filtered_venue_offer(self, client):
-        venue, _ = utils.create_offerer_provider_linked_to_venue()
-        offer = offers_factories.ThingOfferFactory(venue=venue)
-        offers_factories.ThingOfferFactory(
-            venue__managingOfferer=venue.managingOfferer
-        )  # offer attached to other venue
-
-        with testing.assert_no_duplicated_queries():
-            response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-                f"/public/offers/v1/products?venueId={venue.id}"
-            )
-
-        assert response.status_code == 200
-        assert response.json["pagination"] == {
-            "currentPage": 1,
-            "itemsCount": 1,
-            "itemsTotal": 1,
-            "lastPage": 1,
-            "limitPerPage": 50,
-            "pagesLinks": {
-                "current": f"http://localhost/public/offers/v1/products?venueId={venue.id}&page=1&limit=50",
-                "first": f"http://localhost/public/offers/v1/products?venueId={venue.id}&page=1&limit=50",
-                "last": f"http://localhost/public/offers/v1/products?venueId={venue.id}&page=1&limit=50",
-                "next": None,
-                "previous": None,
-            },
-        }
-        assert [product["id"] for product in response.json["products"]] == [offer.id]
-
-    def test_404_when_venue_id_not_tied_to_api_key(self, client):
-        offerers_factories.ApiKeyFactory()
-        unrelated_venue = offerers_factories.VenueFactory()
-
-        response = client.with_explicit_token(offerers_factories.DEFAULT_CLEAR_API_KEY).get(
-            f"/public/offers/v1/products?venueId={unrelated_venue.id}"
-        )
-
-        assert response.status_code == 404
-
-
 @pytest.mark.usefixtures("db_session")
 class GetEventsTest:
     ENDPOINT_URL = "http://localhost/public/offers/v1/events"
